+++ randomizer-core/src/main/java/de/alpharogroup/random/RandomExtensions.java
@@ -26,9 +26,8 @@ package de.alpharogroup.random;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
+import java.math.RoundingMode;
 import java.nio.charset.Charset;
-import java.security.NoSuchAlgorithmException;
-import java.security.NoSuchProviderException;
 import java.security.SecureRandom;
 import java.text.DecimalFormat;
 import java.text.NumberFormat;
@@ -59,16 +58,10 @@ public final class RandomExtensions
 
 	/** The secure random. */
 	private static SecureRandom secureRandom;
+
 	static
 	{
-		try
-		{
-			secureRandom = SecureRandomBean.builder().build();
-		}
-		catch (NoSuchAlgorithmException | NoSuchProviderException e)
-		{
-			throw new RuntimeException("initialization of SecureRandom failed.", e);
-		}
+		secureRandom = SecureRandomFactory.newSecureRandom();
 	}
 
 	/**
@@ -924,5 +917,26 @@ public final class RandomExtensions
 			.getBytes(charset);
 	}
 
+	/**
+	 * Generates a random {@link BigInteger}
+	 *
+	 * @return the random {@link BigInteger}
+	 */
+	public static BigInteger randomBigInteger()
+	{
+		return new BigInteger(randomInt(180), secureRandom);
+	}
+
+	/**
+	 * Generates a random {@link BigDecimal}
+	 *
+	 * @return the random {@link BigDecimal}
+	 */
+	public static BigDecimal randomBigDecimal()
+	{
+		BigDecimal bigDecimal = new BigDecimal(randomDouble());
+		bigDecimal.setScale(randomInt(2), RoundingMode.HALF_DOWN);
+		return bigDecimal;
+	}
 
 }