+++ randomizer-core/src/test/java/de/alpharogroup/random/RandomExtensionsTest.java
@@ -43,6 +43,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.UUID;
 
 import org.apache.commons.lang3.ArrayUtils;
 import org.meanbean.test.BeanTestException;
@@ -326,6 +327,17 @@ public class RandomExtensionsTest extends BaseTestCase
 	}
 
 	/**
+	 * Test method for {@link RandomExtensions#getRandomString(int, int)}
+	 */
+	@Test(enabled = true)
+	public void testGetRandomStringwithStartEnd()
+	{
+		String randomString = RandomExtensions.getRandomString(3, 25);
+		assertNotNull(randomString);
+		assertTrue(MathExtensions.isBetween(3, 25, randomString.length(), true, true));
+	}
+
+	/**
 	 * Test method for {@link RandomExtensions#newRandomPixel()}
 	 */
 	@Test(enabled = true)
@@ -355,6 +367,26 @@ public class RandomExtensionsTest extends BaseTestCase
 	{
 		byte[] newSalt = RandomExtensions.newSalt();
 		assertNotNull(newSalt);
+	}
+
+	/**
+	 * Test method for {@link RandomExtensions#randomBigInteger()}
+	 */
+	@Test
+	public void testRandomBigDecimal()
+	{
+		BigDecimal randomBigDecimal = RandomExtensions.randomBigDecimal();
+		assertNotNull(randomBigDecimal);
+	}
+
+	/**
+	 * Test method for {@link RandomExtensions#randomBigInteger()}
+	 */
+	@Test
+	public void testRandomBigInteger()
+	{
+		BigInteger randomBigInteger = RandomExtensions.randomBigInteger();
+		assertNotNull(randomBigInteger);
 	}
 
 	/**
@@ -397,13 +429,12 @@ public class RandomExtensionsTest extends BaseTestCase
 	}
 
 	/**
-	 * Test method for {@link RandomExtensions#randomDouble(double)}
+	 * Test method for {@link RandomExtensions#randomDouble()}
 	 */
 	@Test(enabled = true)
 	public void testRandomDouble()
 	{
-		double random = RandomExtensions
-			.randomDouble(RandomExtensions.randomDoubleBetween(0.0, 10.0));
+		double random = RandomExtensions.randomDouble();
 		assertTrue(MathExtensions.isBetween(Double.MIN_VALUE, Double.MAX_VALUE, random));
 	}
 
@@ -430,6 +461,17 @@ public class RandomExtensionsTest extends BaseTestCase
 	}
 
 	/**
+	 * Test method for {@link RandomExtensions#randomDouble(double)}
+	 */
+	@Test(enabled = true)
+	public void testRandomDoubleDouble()
+	{
+		double random = RandomExtensions
+			.randomDouble(RandomExtensions.randomDoubleBetween(0.0, 10.0));
+		assertTrue(MathExtensions.isBetween(Double.MIN_VALUE, Double.MAX_VALUE, random));
+	}
+
+	/**
 	 * Test method for {@link RandomExtensions#randomFloat()}
 	 */
 	@Test(enabled = true)
@@ -438,6 +480,7 @@ public class RandomExtensionsTest extends BaseTestCase
 		float random = RandomExtensions.randomFloat();
 		assertTrue(MathExtensions.isBetween(Float.MIN_VALUE, Float.MAX_VALUE, random));
 	}
+
 
 	/**
 	 * Test method for {@link RandomExtensions#randomFloatBetween(float, float)}
@@ -461,7 +504,6 @@ public class RandomExtensionsTest extends BaseTestCase
 		float random = RandomExtensions.randomFloatBetween(0.0f, 10.0f, "###,###.###");
 		assertTrue(MathExtensions.isBetween(0.0f, 10.0f, random));
 	}
-
 
 	/**
 	 * Test method for {@link RandomExtensions#randomFloat(float)}
@@ -580,8 +622,24 @@ public class RandomExtensionsTest extends BaseTestCase
 	@Test(enabled = true)
 	public void testRandomSerialNumber()
 	{
-		BigInteger randomSerialNumber = RandomExtensions.randomSerialNumber();
-		assertNotNull(randomSerialNumber);
+		for (int i = 0; i < 10; i++)
+		{
+			BigInteger randomSerialNumber = RandomExtensions.randomSerialNumber();
+			assertNotNull(randomSerialNumber);
+		}
+	}
+
+	/**
+	 * Test method for {@link RandomExtensions#randomShort()}
+	 */
+	@Test(enabled = true)
+	public void testRandomShort()
+	{
+		for (int i = 0; i < 10; i++)
+		{
+			short randomShort = RandomExtensions.randomShort();
+			assertTrue(MathExtensions.isBetween(-32768, 32767, randomShort, true, true));
+		}
 	}
 
 	/**
@@ -630,6 +688,16 @@ public class RandomExtensionsTest extends BaseTestCase
 	{
 		final String randomToken = RandomExtensions.randomToken();
 		assertNotNull(randomToken);
+	}
+
+	/**
+	 * Test method for {@link RandomExtensions#randomUUID()}
+	 */
+	@Test
+	public void testRandomUUID()
+	{
+		UUID randomUUID = RandomExtensions.randomUUID();
+		assertNotNull(randomUUID);
 	}
 
 	/**